Major Refactoring of the schema documentation generator.
