+++ transform/src/schemas/doc/generate_markdown.py
@@ -69,57 +69,4 @@ class MarkdownGenerator(object):
             with open(filename, 'w') as f:
                 f.write(retStr)
             return retStr
-    
-class PyMDGenerator(MarkdownGenerator):
-    def __init__(self, database: JsObject):
-        super().__init__(database)
-
-    def _generate_docstring(self, docstring: str):
-        """Generate markdown for a single column."""
-        if (docstring is None):
-            self.add_line(1, "No documentation provided.")
-            return
-
-        split_doc_string = docstring.split("\n")
-        if len(split_doc_string) == 1:
-            self.add_line(1, f'{docstring}')
-            return
-        
-        for line in split_doc_string:
-            self.add_line(1, f'{line}')
-
-        
-
-    def generate_column_markdown(self, indent: int, column: JsObject):
-        """Generate markdown for a single column."""       
-        style = "columnstyle" if not column["primary_key"] else "keycolumnstyle"
-        self.add_line(0, f'??? {style} "{column["name"] + "    (" +column["type"]})"')
-        self.add_line(1, "```{.codeblock}")
-        self._generate_docstring(column["doc"])
-        #self.add_line(indent+1, "")
-        self.add_line(1, "```")
-        self.add_line(indent, f'\n')
-    
-    
-    def generate_table_markdown(self, indent: int, table: JsObject):
-        """Generate markdown for a single table."""
-        self.add_line(indent, f'### {table["table_name"]}\n')
-        for column in table["columns"]:
-            self.generate_column_markdown(indent+1,column)
-        self.add_line(indent, f'\n')
-
-    def generate_schema_markdown(self, indent: int, schema: JsObject):
-        """Generate markdown for all tables in the given schema."""
-        self.add_line(indent, f'## {schema["schema_name"]}')
-        for table in schema["tables"]:
-            self.generate_table_markdown(indent, table)
-        self.add_line(indent, f'\n')
-
-    def generate_database_markdown(self, database: JsObject):
-        """Generate markdown for all schemas in the given database."""
-        self.add_line(0, f'# {database["database_name"]} Schemas')
-        self.add_line(0, f'[TOC]')
-
-        for schema in database["schemas"].values():
-            self.generate_schema_markdown(0, schema)
